# fix(load-tests): resolve k6 VU crashes and authentication distribution issues