@AutoGPT-Agent 2025-09-22
Thank you for the detailed PR description that clearly explains the problem, solution, and impact. The breakdown of issues and improvements is particularly helpful.

A few observations:

1. Your PR is missing the standard checklist that's required in our PR template. While this may be acceptable for non-material changes, it would be helpful to include it or explicitly note why it's omitted.

2. Your PR title follows our conventional commit format, which is great.

3. Unfortunately, without seeing the actual code changes (the diff is empty in what I received), I cannot fully evaluate whether the changes match the scope described or if there are any specific code-related concerns.

Could you please:
- Add the standard checklist or explain why it's omitted
- Ensure that at least 80% of your changes directly relate to fixing the k6 load test issues mentioned in the title

Once these items are addressed, or if they're already taken care of in the actual PR, this should be ready for further review.